working on finalizing the triangle fill

+++ point2d.h
@@ -3,8 +3,8 @@
 
 struct Point2D
 {
-	int x{};
-	int y{};
+	float x{};
+	float y{};
 };
 
 #endif
